requirements.txt
# Stinger Test Stand Requirements
# Python 3.10+

# UI Framework
PyQt6>=6.6.0

# Hardware
labjack-ljm>=1.0.0
pyserial>=3.5

# Database
SQLAlchemy>=2.0.0
pyodbc>=5.0.0

# Configuration
PyYAML>=6.0

# State Machine
transitions>=0.9.0

# Utilities
numpy>=1.24.0

# Visualization
pyqtgraph>=0.13.0

# Logging (stdlib, but listed for documentation)
# logging - included in Python stdlib

# Development / Testing
pytest>=7.0.0
pytest-qt>=4.2.0
pytest-cov>=5.0.0
